Add new data data structure for holding tooltip states
This begins a series of changes to add support for two tooltips:
one sticky and one hovering. The user can click on a line chart
to create a sticky tooltip and then continue to hover in other
places on the chart to see a tooltip as they hover.

We use self.tooltips as a map to hold two named states:
STICKY_TOOLTIP and HOVER_TOOLTIP. We use these to store the node
and position data.

For STICKY_TOOLTIP, we also hold data about the sticky data point
under the `item` key. We use this to compare datapoints when hovering
so we don't show a tooltip on top of the sticky one. `item` also has
data used to highlight the series itself.