Implement new hover and click logic
Note that hovering creates a new nod and attaches it and stickying
a tooltip moves it in state but resuses the node for increased perf.

Supported uses include:
- Hovering on a chart
- Clicking to sticky a hovered tooltip
- Seeing a hovered tooltip while one is sticky
- Closing a locked tooltip by clicking on the data point it points at
- Closing a locked tooltip by clicking on an empty part of the chart
- Repeatedly clicking on a data point to open/close a tooltip

We made the decision to note show a hover tooltip over a sticky one
if hovering over the same data point because it is redundant.

Note each chart has its own instance of this plugin but some things
are shared between chart (the dot on the graph, the vertical line).